Allow at_css/at_xml to return multiple values
If multiple nodes match the selector, get all of them, and join with a
semicolon (which CSV-to-Popolo knows how to handle).

We should probably really be able to return this directly as a list, and
have the scraper know how to flatten that to something suitable for
ScraperWiki, but that can be a later step.